+++ bitcoin/src/base58.rs
@@ -6,7 +6,6 @@
 //! strings respectively.
 //!
 
-use core::convert::TryInto;
 use core::{fmt, iter, slice, str};
 
 use hashes::{sha256d, Hash};