+++ bitcoin/src/parse.rs
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: CC0-1.0
 
-use core::convert::TryFrom;
 use core::fmt;
 use core::str::FromStr;