+++ bitcoin/src/psbt/map/output.rs
@@ -1,7 +1,5 @@
 // SPDX-License-Identifier: CC0-1.0
 
-use core::convert::TryFrom;
-
 use secp256k1::XOnlyPublicKey;
 use {core, secp256k1};