+++ bitcoin/tests/psbt.rs
@@ -1,7 +1,6 @@
 //! Tests PSBT integration vectors from BIP 174
 //! defined at <https://github.com/bitcoin/bips/blob/master/bip-0174.mediawiki#test-vectors>
 
-use core::convert::TryFrom;
 use std::collections::BTreeMap;
 use std::str::FromStr;