+++ contrib/test.sh
@@ -4,22 +4,6 @@ set -ex
 
 CRATES="bitcoin hashes internals fuzz"
 DEPS="recent minimal"
-MSRV="1\.48\.0"
-
-# Test pinned versions.
-if cargo --version | grep ${MSRV}; then
-    cargo update -p serde_json --precise 1.0.99
-    cargo update -p serde --precise 1.0.156
-    cargo update -p quote --precise 1.0.30
-    cargo update -p proc-macro2 --precise 1.0.63
-    cargo update -p serde_test --precise 1.0.175
-    cargo update -p schemars --precise 0.8.12
-    # byteorder 1.5.0 uses edition 2021
-    cargo update -p byteorder --precise 1.4.3
-
-    # Build MSRV with pinned versions.
-    cargo check --all-features --all-targets
-fi
 
 for dep in $DEPS
 do