+++ bitcoin/CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Bump MSRV to Rust 1.56.1
+
 # 0.31.1 - 2023-10-18
 
 - Bump MSRV to Rust 1.48.0 [#1729](https://github.com/rust-bitcoin/rust-bitcoin/pull/1729)